requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
fastapi==0.110.0
ffmpeg-python==0.2.0
Flask==3.0.2
future==1.0.0
h11==0.14.0
httptools==0.6.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
numpy==1.26.4
pillow==10.2.0
proglog==0.1.10
PyAutoGUI==0.9.54
pydantic==2.6.3
pydantic_core==2.16.3
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
pytube==15.0.0
pytweening==1.2.0
pywhatkit==5.4
PyYAML==6.0.1
requests==2.31.0
setuptools==69.1.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wikipedia==1.4.0
